+++ folio-secret-store/folio-secret-store-common/src/main/java/org/folio/tools/store/utils/TlsProperties.java
@@ -1,0 +1,20 @@
+package org.folio.tools.store.utils;
+
+import static org.apache.commons.lang3.StringUtils.defaultString;
+
+import lombok.Builder;
+import lombok.Value;
+
+@Value
+@Builder
+public class TlsProperties {
+
+  boolean enabled;
+  String trustStorePath;
+  String trustStorePassword;
+  String trustStoreType;
+
+  public String getTrustStorePassword() {
+    return defaultString(trustStorePassword);
+  }
+}