+++ folio-secret-store/folio-secret-store-starter/src/main/java/org/folio/tools/store/properties/FsspProperties.java
@@ -1,0 +1,18 @@
+package org.folio.tools.store.properties;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import org.hibernate.validator.constraints.URL;
+import org.springframework.validation.annotation.Validated;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+@Validated
+public class FsspProperties extends FsspConfigProperties {
+
+  @URL
+  private String address;
+  @NotNull
+  private Boolean enableSsl;
+}